packages/protocols/x402-libertai/examples/agentkit/python/requirements.txt
coinbase-agentkit>=0.7.0
libertai-agentkit-plugin>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
